Fixed sorting on all pages
The reviews are now in highest to lowest and the restaurants are sorted from highest rating to lowest rating

+++ DSFinalProjectFXML/src/detail/DetailController.java
@@ -99,7 +99,7 @@ public class DetailController implements Initializable{
     	reviewHeader = new HBox(reviewTitle, addReview);
     	menuList = new ListView<FoodItem>();
     	reviewList = new ListView<Rating>();
-    	switch(restaurant.getAvgRating()) {
+    	switch((int)restaurant.getAvgRating()) {
         case 1:
         	rating.setImage(new Image("/starIcons/1star.png"));
         	break;
